feat(firmware): full-duplex 32-bit I2S, mic capture proven on device (HW-B2)
XVF3800 INT variant requires 32-bit I2S slots (XMOS Programming Guide
Table 2.1; Seeed reference uses AUDIO_I2S_BITS=32). Driver now creates
a TX+RX channel pair sharing BCLK/WS with int32_t buffers.

Diagnostics that isolated the fault: line-ownership probe proved the
XMOS is a pure I2S slave (0% duty when idle); pull-down test exposed
floating-pin crosstalk as the source of garbage reads; mute ruled out
as gate. After 32-bit switch: user speech/claps captured with clear
SNR contrast, L=processed beam R=raw mic per XMOS mux defaults.

Tick HW-B2 both directions. Update docs and project state.

+++ firmware/components/pericles_core/smoke_test.c
@@ -37,7 +37,7 @@ smoke_result_t smoke_test_audio_i2s(void) {
     xvf_err_t err = xvf3800_i2s_init();
     if (err != XVF_OK) return SMOKE_FAIL;
 
-    int16_t silence[256] = { 0 };
+    int32_t silence[256] = { 0 };
     err = xvf3800_i2s_write(silence, 256);
 
     xvf3800_i2s_deinit();
